Policies added

+++ Alumni/app/Http/Controllers/ProfileController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Intervention\Image\Facades\Image;
 use App\Models\Profile;
 
+
 class ProfileController extends Controller
 {
     /**
@@ -43,7 +44,10 @@ class ProfileController extends Controller
     public function show($id)
     {
         //show a profile
-        return Profile::find($id);
+        $profile = Profile::find($id);
+        //check policy first
+        // $this->authorize('view', $profile);
+        return $profile;
     }
 
     /**
@@ -55,8 +59,14 @@ class ProfileController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //update a profile
+        
+        //Get profile id
         $profile = Profile::find($id);
+        //check policy first
+        $this->authorize('update', $profile);
+        //update profile
+        $profile->update($request->all());
+
         $validatedDataProfile = request()->validate([
             //'user_id',
             'phone' => '',
@@ -97,6 +107,10 @@ class ProfileController extends Controller
      */
     public function destroy($id)
     {
+        //Get profile id
+        $profile = Profile::find($id);
+        //check policy first
+        $this->authorize('delete', $profile);
         //delete profile
         return Profile::destroy($id);
     }
